Implement first 4 steps of the suggested fix for Issue 5849: * Remove all token* constructors in services.constructor package * Remove HTS related table (token, token_transfer, nft_transfer) query and processing in persistence.TransactionRepository * Remove fungible / non fungible token account balance support in persistence.AccountRepository * Remove HTS BDD test case
Also, for the first three bullets, fix the tests broken by removing that code.
