+++ hedera-mirror-rosetta/app/persistence/account_test.go
@@ -421,262 +421,6 @@ func (suite *accountRepositorySuite) TestGetAccountIdDbConnectionError() {
 	assert.Equal(suite.T(), types.AccountId{}, actual)
 }
 
-func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlock() {
-	// given
-	// tokens created at or before first account balance snapshot will not show up in account balance response
-	// transfers before or at the snapshot timestamp should not affect balance calculation
-	accountId := suite.accountId
-	repo := NewAccountRepository(dbClient)
-
-	hbarAmount := &types.HbarAmount{Value: initialAccountBalance + sum(cryptoTransferAmounts)}
-	token1Amount := types.NewTokenAmount(token1, sum(token1TransferAmounts[:2]))
-	token2Amount := types.NewTokenAmount(token2, sum(token2TransferAmounts[:2]))
-	token3Amount := types.NewTokenAmount(token3, 2)
-	token4Amount := types.NewTokenAmount(token4, 0)
-	expectedAmounts := types.AmountSlice{hbarAmount, token1Amount, token2Amount, token3Amount, token4Amount}
-
-	// when
-	// query
-	actualAmounts, accountIdString, err := repo.RetrieveBalanceAtBlock(defaultContext, accountId, consensusTimestamp)
-
-	// then
-	assert.Nil(suite.T(), err)
-	assert.Equal(suite.T(), suite.accountIdString, accountIdString)
-	assert.ElementsMatch(suite.T(), expectedAmounts, actualAmounts)
-
-	// when
-	// query at dissociateTimestamp, balances for token2 and token3 should be 0
-	actualAmounts, accountIdString, err = repo.RetrieveBalanceAtBlock(defaultContext, accountId, dissociateTimestamp)
-	token2Amount = types.NewTokenAmount(token2, 0)
-	token3Amount = types.NewTokenAmount(token3, 0)
-	expectedAmounts = types.AmountSlice{hbarAmount, token1Amount, token2Amount, token3Amount, token4Amount}
-
-	// then
-	assert.Nil(suite.T(), err)
-	assert.Equal(suite.T(), suite.accountIdString, accountIdString)
-	assert.ElementsMatch(suite.T(), expectedAmounts, actualAmounts)
-}
-
-func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlockAfterSecondSnapshot() {
-	// given
-	// remove any transfers in db. with the balance info in the second snapshot, this test verifies the account balance
-	// is computed with additional transfers applied on top of the snapshot
-	accountId := suite.accountId
-	db.ExecSql(dbClient, truncateCryptoTransferFileSql, truncateNftTransferSql, truncateTokenTransferSql)
-	tdomain.NewAccountBalanceFileBuilder(dbClient, secondSnapshotTimestamp).
-		AddAccountBalance(account1, initialAccountBalance+sum(cryptoTransferAmounts)).
-		AddTokenBalance(account1, encodedTokenId1, sum(token1TransferAmounts[:2])).
-		AddTokenBalance(account1, encodedTokenId2, sum(token2TransferAmounts[:2])).
-		AddTokenBalance(account1, encodedTokenId3, 2).
-		AddTokenBalance(account1, encodedTokenId4, 0).
-		TimeOffset(-1).
-		Persist()
-	// extra transfers
-	// account balance file time offset is -1, which means the transfer at secondSnapshotTimestamp is not included in
-	// the snapshot. The balance response should include the amount in this transfer too.
-	tdomain.NewTokenTransferBuilder(dbClient).
-		AccountId(account1).
-		Amount(10).
-		Timestamp(secondSnapshotTimestamp).
-		TokenId(encodedTokenId1).
-		Persist()
-	tdomain.NewTokenTransferBuilder(dbClient).
-		AccountId(account1).
-		Amount(12).
-		Timestamp(secondSnapshotTimestamp + 3).
-		TokenId(encodedTokenId2).
-		Persist()
-	tdomain.NewNftTransferBuilder(dbClient).
-		ReceiverAccountId(account1).
-		SenderAccountId(treasury).
-		SerialNumber(6).
-		Timestamp(secondSnapshotTimestamp + 4).
-		TokenId(encodedTokenId3).
-		Persist()
-	// add a self transfer
-	tdomain.NewNftTransferBuilder(dbClient).
-		ReceiverAccountId(account1).
-		SenderAccountId(account1).
-		SerialNumber(6).
-		Timestamp(secondSnapshotTimestamp + 5).
-		TokenId(encodedTokenId3).
-		Persist()
-	hbarAmount := &types.HbarAmount{Value: initialAccountBalance + sum(cryptoTransferAmounts)}
-	token1Amount := types.NewTokenAmount(token1, sum(token1TransferAmounts[:2])+10)
-	token2Amount := types.NewTokenAmount(token2, sum(token2TransferAmounts[:2])+12)
-	token3Amount := types.NewTokenAmount(token3, 3)
-	token4Amount := types.NewTokenAmount(token4, 0)
-	expectedAmount := types.AmountSlice{hbarAmount, token1Amount, token2Amount, token3Amount, token4Amount}
-	repo := NewAccountRepository(dbClient)
-
-	// when
-	actualAmounts, accountIdString, err := repo.RetrieveBalanceAtBlock(
-		defaultContext,
-		accountId,
-		secondSnapshotTimestamp+6,
-	)
-
-	// then
-	assert.Nil(suite.T(), err)
-	assert.Equal(suite.T(), suite.accountIdString, accountIdString)
-	assert.ElementsMatch(suite.T(), expectedAmount, actualAmounts)
-}
-
-func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlockForDeletedAccount() {
-	// given
-	accountId := suite.accountId
-	tdomain.NewEntityBuilder(dbClient, account1, 1, domain.EntityTypeAccount).
-		Deleted(true).
-		ModifiedTimestamp(accountDeleteTimestamp).
-		Persist()
-	expectedAmounts := types.AmountSlice{
-		&types.HbarAmount{},
-		types.NewTokenAmount(token1, 0),
-		types.NewTokenAmount(token2, 0),
-		types.NewTokenAmount(token3, 0),
-		types.NewTokenAmount(token4, 0),
-	}
-	repo := NewAccountRepository(dbClient)
-
-	// when
-	// account is deleted before the third account balance file, so there is no balance info in the file. querying the
-	// account balance for a timestamp after the third account balance file should then return the balance at the time
-	// the account is deleted
-	actualAmounts, accountIdString, err := repo.RetrieveBalanceAtBlock(
-		defaultContext,
-		accountId,
-		thirdSnapshotTimestamp+10,
-	)
-
-	// then
-	assert.Nil(suite.T(), err)
-	assert.Equal(suite.T(), suite.accountIdString, accountIdString)
-	assert.ElementsMatch(suite.T(), expectedAmounts, actualAmounts)
-}
-
-func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlockAtAccountDeletionTime() {
-	// given
-	accountId := suite.accountId
-	tdomain.NewEntityBuilder(dbClient, account1, 1, domain.EntityTypeAccount).
-		Deleted(true).
-		ModifiedTimestamp(accountDeleteTimestamp).
-		Persist()
-	expectedAmounts := types.AmountSlice{
-		&types.HbarAmount{},
-		types.NewTokenAmount(token1, 0),
-		types.NewTokenAmount(token2, 0),
-		types.NewTokenAmount(token3, 0),
-		types.NewTokenAmount(token4, 0),
-	}
-	repo := NewAccountRepository(dbClient)
-
-	// when
-	actualAmounts, accountIdString, err := repo.RetrieveBalanceAtBlock(
-		defaultContext,
-		accountId,
-		accountDeleteTimestamp,
-	)
-
-	// then
-	assert.Nil(suite.T(), err)
-	assert.Equal(suite.T(), suite.accountIdString, accountIdString)
-	assert.ElementsMatch(suite.T(), expectedAmounts, actualAmounts)
-}
-
-func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlockNoAccountEntity() {
-	// given
-	accountId := suite.accountId
-	db.ExecSql(dbClient, truncateEntitySql)
-	hbarAmount := &types.HbarAmount{Value: initialAccountBalance + sum(cryptoTransferAmounts)}
-	token1Amount := types.NewTokenAmount(token1, sum(token1TransferAmounts[:2]))
-	token2Amount := types.NewTokenAmount(token2, sum(token2TransferAmounts[:2]))
-	token3Amount := types.NewTokenAmount(token3, 2)
-	token4Amount := types.NewTokenAmount(token4, 0)
-	expectedAmounts := types.AmountSlice{hbarAmount, token1Amount, token2Amount, token3Amount, token4Amount}
-	repo := NewAccountRepository(dbClient)
-
-	// when
-	actualAmounts, accountIdString, err := repo.RetrieveBalanceAtBlock(
-		defaultContext,
-		accountId,
-		consensusTimestamp,
-	)
-
-	// then
-	assert.Nil(suite.T(), err)
-	assert.Empty(suite.T(), accountIdString)
-	assert.ElementsMatch(suite.T(), expectedAmounts, actualAmounts)
-}
-
-func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlockNoTokenEntity() {
-	// given
-	accountId := suite.accountId
-	db.ExecSql(dbClient, truncateTokenSql)
-	repo := NewAccountRepository(dbClient)
-
-	// no token entities, so only hbar balance
-	hbarAmount := &types.HbarAmount{Value: initialAccountBalance + sum(cryptoTransferAmounts)}
-	expectedAmounts := types.AmountSlice{hbarAmount}
-
-	// when
-	actualAmounts, accountIdString, err := repo.RetrieveBalanceAtBlock(
-		defaultContext,
-		accountId,
-		consensusTimestamp,
-	)
-
-	// then
-	assert.Nil(suite.T(), err)
-	assert.Equal(suite.T(), suite.accountIdString, accountIdString)
-	assert.ElementsMatch(suite.T(), expectedAmounts, actualAmounts)
-}
-
-func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlockNoInitialBalance() {
-	// given
-	accountId := suite.accountId
-	db.ExecSql(dbClient, truncateAccountBalanceSql, truncateTokenBalanceSql)
-
-	hbarAmount := &types.HbarAmount{Value: sum(cryptoTransferAmounts)}
-	token1Amount := types.NewTokenAmount(token1, sum(token1TransferAmounts[:2]))
-	token2Amount := types.NewTokenAmount(token2, sum(token2TransferAmounts[:2]))
-	token3Amount := types.NewTokenAmount(token3, 2)
-	token4Amount := types.NewTokenAmount(token4, 0)
-	expectedAmounts := types.AmountSlice{hbarAmount, token1Amount, token2Amount, token3Amount, token4Amount}
-
-	repo := NewAccountRepository(dbClient)
-
-	// when
-	actualAmounts, accountIdString, err := repo.RetrieveBalanceAtBlock(
-		defaultContext,
-		accountId,
-		consensusTimestamp,
-	)
-
-	// then
-	assert.Nil(suite.T(), err)
-	assert.Equal(suite.T(), suite.accountIdString, accountIdString)
-	assert.ElementsMatch(suite.T(), expectedAmounts, actualAmounts)
-}
-
-func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlockNoAccountBalanceFile() {
-	// given
-	db.ExecSql(dbClient, truncateAccountBalanceFileSql)
-	accountId := suite.accountId
-	repo := NewAccountRepository(dbClient)
-
-	// when
-	actualAmounts, accountIdString, err := repo.RetrieveBalanceAtBlock(
-		defaultContext,
-		accountId,
-		consensusTimestamp,
-	)
-
-	// then
-	assert.NotNil(suite.T(), err)
-	assert.Empty(suite.T(), accountIdString)
-	assert.Nil(suite.T(), actualAmounts)
-}
-
 func (suite *accountRepositorySuite) TestRetrieveBalanceAtBlockDbConnectionError() {
 	// given
 	accountId := suite.accountId
