+++ hedera-mirror-rosetta/app/services/construction/crypto_transfer_transaction_constructor_test.go
@@ -41,18 +41,11 @@ var (
 	currencyHbar = types.CurrencyHbar
 
 	defaultSerialNumbers = []int64{1}
-	defaultSigners       = []types.AccountId{accountIdA, accountIdB}
+	defaultSigners       = []types.AccountId{accountIdA}
 	defaultTransfers     = []transferOperation{
 		{accountId: accountIdA, amount: &types.HbarAmount{Value: -15}},
 		{accountId: accountIdB, amount: &types.HbarAmount{Value: 15}},
-		{accountId: accountIdB, amount: types.NewTokenAmount(dbTokenA, -25)},
-		{accountId: accountIdA, amount: types.NewTokenAmount(dbTokenA, 25)},
-		{accountId: accountIdA, amount: types.NewTokenAmount(dbTokenB, -30)},
-		{accountId: accountIdB, amount: types.NewTokenAmount(dbTokenB, 30)},
-		{accountId: accountIdA, amount: types.NewTokenAmount(dbTokenC, -1).SetSerialNumbers(defaultSerialNumbers)},
-		{accountId: accountIdB, amount: types.NewTokenAmount(dbTokenC, 1).SetSerialNumbers(defaultSerialNumbers)},
 	}
-	nftId               = hedera.NftID{TokenID: tokenIdC, SerialNumber: 1}
 	outOfRangeAccountId = hedera.AccountID{Shard: 2 << 15, Realm: 2 << 16, Account: 2<<32 + 5}
 	outOfRangeTokenId   = hedera.TokenID{Shard: 2 << 15, Realm: 2 << 16, Token: 2 << 32}
 )
@@ -129,23 +122,12 @@ func (suite *cryptoTransferTransactionConstructorSuite) TestParse() {
 		return hedera.NewTransferTransaction().
 			AddHbarTransfer(sdkAccountIdA, hedera.HbarFromTinybar(-15)).
 			AddHbarTransfer(sdkAccountIdB, hedera.HbarFromTinybar(15)).
-			AddTokenTransferWithDecimals(tokenIdA, sdkAccountIdA, -25, uint32(dbTokenA.Decimals)).
-			AddTokenTransferWithDecimals(tokenIdA, sdkAccountIdB, 25, uint32(dbTokenA.Decimals)).
-			AddTokenTransferWithDecimals(tokenIdB, sdkAccountIdB, -35, uint32(dbTokenB.Decimals)).
-			AddTokenTransferWithDecimals(tokenIdB, sdkAccountIdA, 35, uint32(dbTokenB.Decimals)).
-			AddNftTransfer(nftId, sdkAccountIdA, sdkAccountIdB).
 			SetTransactionID(hedera.TransactionIDGenerate(sdkAccountIdA))
 	}
 
 	expectedTransfers := []string{
 		transferStringify(sdkAccountIdA, -15, currencyHbar.Symbol, uint32(currencyHbar.Decimals), 0),
 		transferStringify(sdkAccountIdB, 15, currencyHbar.Symbol, uint32(currencyHbar.Decimals), 0),
-		transferStringify(sdkAccountIdA, -25, tokenIdA.String(), uint32(dbTokenA.Decimals), 0),
-		transferStringify(sdkAccountIdB, 25, tokenIdA.String(), uint32(dbTokenA.Decimals), 0),
-		transferStringify(sdkAccountIdB, -35, tokenIdB.String(), uint32(dbTokenB.Decimals), 0),
-		transferStringify(sdkAccountIdA, 35, tokenIdB.String(), uint32(dbTokenB.Decimals), 0),
-		transferStringify(sdkAccountIdA, -1, tokenIdC.String(), 0, 1), // nft
-		transferStringify(sdkAccountIdB, 1, tokenIdC.String(), 0, 1),  // nft
 	}
 
 	var tests = []struct {
@@ -165,63 +147,11 @@ func (suite *cryptoTransferTransactionConstructorSuite) TestParse() {
 			expectError: true,
 		},
 		{
-			name: "NoExpectedDecimals",
-			getTransaction: func() interfaces.Transaction {
-				return hedera.NewTransferTransaction().
-					AddTokenTransfer(tokenIdA, sdkAccountIdA, -25).
-					AddTokenTransfer(tokenIdA, sdkAccountIdB, 25).
-					SetTransactionID(hedera.TransactionIDGenerate(sdkAccountIdA))
-			},
-			expectError: true,
-		},
-		{
-			name: "InvalidFungibleCommonTokenId",
-			getTransaction: func() interfaces.Transaction {
-				return hedera.NewTransferTransaction().
-					AddTokenTransferWithDecimals(outOfRangeTokenId, sdkAccountIdA, -25, decimals).
-					AddTokenTransferWithDecimals(outOfRangeTokenId, sdkAccountIdB, 25, decimals).
-					SetTransactionID(hedera.TransactionIDGenerate(sdkAccountIdA))
-			},
-			expectError: true,
-		},
-		{
-			name: "InvalidNonFungibleUniqueTokenId",
-			getTransaction: func() interfaces.Transaction {
-				return hedera.NewTransferTransaction().
-					AddNftTransfer(
-						hedera.NftID{TokenID: outOfRangeTokenId, SerialNumber: 1},
-						sdkAccountIdA,
-						sdkAccountIdB,
-					).
-					SetTransactionID(hedera.TransactionIDGenerate(sdkAccountIdA))
-			},
-			expectError: true,
-		},
-		{
 			name: "OutOfRangeAccountIdHbarTransfer",
 			getTransaction: func() interfaces.Transaction {
 				return hedera.NewTransferTransaction().
 					AddHbarTransfer(outOfRangeAccountId, hedera.HbarFromTinybar(-15)).
 					AddHbarTransfer(sdkAccountIdB, hedera.HbarFromTinybar(15)).
-					SetTransactionID(hedera.TransactionIDGenerate(outOfRangeAccountId))
-			},
-			expectError: true,
-		},
-		{
-			name: "OutOfRangeAccountIdFungibleTokenTransfer",
-			getTransaction: func() interfaces.Transaction {
-				return hedera.NewTransferTransaction().
-					AddTokenTransferWithDecimals(tokenIdA, outOfRangeAccountId, -25, uint32(dbTokenA.Decimals)).
-					AddTokenTransferWithDecimals(tokenIdA, sdkAccountIdB, 25, uint32(dbTokenA.Decimals)).
-					SetTransactionID(hedera.TransactionIDGenerate(outOfRangeAccountId))
-			},
-			expectError: true,
-		},
-		{
-			name: "OutOfRangeAccountIdNonFungibleTokenTransfer",
-			getTransaction: func() interfaces.Transaction {
-				return hedera.NewTransferTransaction().
-					AddNftTransfer(nftId, outOfRangeAccountId, sdkAccountIdB).
 					SetTransactionID(hedera.TransactionIDGenerate(outOfRangeAccountId))
 			},
 			expectError: true,
@@ -287,39 +217,6 @@ func (suite *cryptoTransferTransactionConstructorSuite) TestPreprocess() {
 			transfers: []transferOperation{
 				{accountId: accountIdA, amount: &types.HbarAmount{Value: -15}},
 				{accountId: accountIdB, amount: &types.HbarAmount{Value: 25}},
-			},
-			expectError: true,
-		},
-		{
-			name: "InvalidTokenSum",
-			transfers: []transferOperation{
-				{accountId: accountIdA, amount: types.NewTokenAmount(dbTokenA, -15)},
-				{accountId: accountIdB, amount: types.NewTokenAmount(dbTokenA, 20)},
-			},
-			expectError: true,
-		},
-		{
-			name: "InvalidNftAmount",
-			transfers: []transferOperation{
-				{accountId: accountIdA, amount: types.NewTokenAmount(dbTokenC, -2).SetSerialNumbers([]int64{1, 2})},
-				{accountId: accountIdB, amount: types.NewTokenAmount(dbTokenC, 2).SetSerialNumbers([]int64{1, 2})},
-			},
-			expectError: true,
-		},
-		{
-			name: "InvalidNftTransferSum",
-			transfers: []transferOperation{
-				{accountId: accountIdA, amount: types.NewTokenAmount(dbTokenC, -1).SetSerialNumbers([]int64{1})},
-			},
-			expectError: true,
-		},
-		{
-			name: "DoubleNftTransfer",
-			transfers: []transferOperation{
-				{accountId: accountIdA, amount: types.NewTokenAmount(dbTokenC, -1).SetSerialNumbers([]int64{1})},
-				{accountId: accountIdA, amount: types.NewTokenAmount(dbTokenC, -1).SetSerialNumbers([]int64{1})},
-				{accountId: accountIdB, amount: types.NewTokenAmount(dbTokenC, 1).SetSerialNumbers([]int64{1})},
-				{accountId: accountIdB, amount: types.NewTokenAmount(dbTokenC, 1).SetSerialNumbers([]int64{1})},
 			},
 			expectError: true,
 		},
@@ -394,9 +291,6 @@ func assertCryptoTransferTransaction(
 	tx, _ := actual.(*hedera.TransferTransaction)
 
 	actualHbarTransfers := tx.GetHbarTransfers()
-	actualTokenTransfers := tx.GetTokenTransfers()
-	actualTokenDecimals := tx.GetTokenIDDecimals()
-	actualNftTransfers := tx.GetNftTransfers()
 
 	actualTransfers := make([]string, 0)
 	for accountId, amount := range actualHbarTransfers {
@@ -404,27 +298,6 @@ func assertCryptoTransferTransaction(
 			actualTransfers,
 			transferStringify(accountId, amount.AsTinybar(), currencyHbar.Symbol, uint32(currencyHbar.Decimals), 0),
 		)
-	}
-
-	for token, tokenTransfers := range actualTokenTransfers {
-		decimals, ok := actualTokenDecimals[token]
-		assert.True(t, ok, "no decimals found for token %s", token.String())
-		for _, transfer := range tokenTransfers {
-			actualTransfers = append(
-				actualTransfers,
-				transferStringify(transfer.AccountID, transfer.Amount, token.String(), decimals, 0),
-			)
-		}
-	}
-
-	for token, nftTransfers := range actualNftTransfers {
-		for _, nftTransfer := range nftTransfers {
-			actualTransfers = append(
-				actualTransfers,
-				transferStringify(nftTransfer.ReceiverAccountID, 1, token.String(), 0, nftTransfer.SerialNumber),
-				transferStringify(nftTransfer.SenderAccountID, -1, token.String(), 0, nftTransfer.SerialNumber),
-			)
-		}
 	}
 
 	assert.ElementsMatch(t, expectedTransfers, actualTransfers)
